Add files for running an infinite loop
Not quite sure where to put this, so putting it here for now

Node.h and Node.cpp make a component type with a single field that upon
receiving an event just sends another basic event back along the link.  Makefile
compiles it into an element library and registers that library. infiniteLoop.py
creates two instances of that component and links them together.

I checked the program on its own and using the SST debugger and it will run as
long as it is let to.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>